move hexgrid construction code to separate file. Remove the redundant data points that causes the table to grow. Give users options to use numba or not, and created empty decorator for properly running the script if not using numba
